requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
python-jose
passlib[bcrypt]
python-multipart